Adding web front end with React, Vite, & FastAPI
Antigravity Prompt: Next let's add a simple local webserver and application to provide a convenient local browser based user experience. We will expose the chat features as well as the health checks and backend, for example a user may with to "upload" an archive vs manually copying it to the working directory. It would be nice to have a stats page to show embedding / tokenization backgrond tasks in progress with estimates completion times.